New changes
Added Freetype texture font rendering. Still broken, need to work on it.
Addded MD5 skeletal animation reading and rendering. Close to porting it
from OpenGL. Changed from FreeTy[e 2.5 to 2.6. Fixed libpng.lib which
was build on 32bit now it is build on 64bit pc. Fixed problems with
assimp. Added md5 files with animation in assets directory

+++ Renderer/MeshLoader/Model.h
@@ -16,7 +16,7 @@ struct TYWRENDERER_API srfTriangles_t
 	drawVert*			verts;			//vert, normal, tex
 
 	int					numIndexes;		//number of indixes
-	uint16_t*			indexes;		//vertex indices
+	uint32_t*			indexes;		//vertex indices
 };
 
 struct TYWRENDERER_API modelSurface_t 
@@ -44,9 +44,9 @@ struct TYWRENDERER_API deformInfo_t
 	drawVert *			verts;
 
 	int					numIndexes;
-	uint16_t *			indexes;
+	uint32_t *			indexes;
 
-	uint16_t *			silIndexes;				// indexes changed to be the first vertex with same XYZ, ignoring normal and texcoords
+	uint32_t*			silIndexes;				// indexes changed to be the first vertex with same XYZ, ignoring normal and texcoords
 
 	int					numMirroredVerts;		// this many verts at the end of the vert list are tangent mirrors
 	int *				mirroredVerts;			// tri->mirroredVerts[0] is the mirror of tri->numVerts - tri->numMirroredVerts + 0
